Initial literal parsing (#4)

+++ model/src/main/scala/org/alax/model/literal.scala
@@ -9,7 +9,7 @@ case class CharacterLiteral(override val value: java.lang.Character) extends bas
 case class IntegerLiteral(override val value: java.lang.Integer) extends base.Literal(value = value);
 
 
-case class FlaotLiteral(override val value: java.lang.Float) extends base.Literal(value = value);
+case class FlaotLiteral(override val value: java.lang.Double) extends base.Literal(value = value);
 
 
 case class StringLiteral(override val value: java.lang.String) extends base.Literal(value = value);
